# docs/VMOGD3swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "VMOGD3Project",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "description": "",
      "name": "token",
      "in": "header"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/admins": {
      "post": {
        "summary": "Add admin",
        "tags": [
          "Admin"
        ],
        "operationId": "Addadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddadminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "List admins",
        "tags": [
          "Admin"
        ],
        "operationId": "Listadmins",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/admins/{id}": {
      "get": {
        "summary": "Get admin by id",
        "tags": [
          "Admin"
        ],
        "operationId": "Getadminbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update admin",
        "tags": [
          "Admin"
        ],
        "operationId": "Updateadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateadminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete admin",
        "tags": [
          "Admin"
        ],
        "operationId": "Deleteadmin",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteadminRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/customers/{id}": {
      "get": {
        "summary": "Get customer by id",
        "tags": [
          "Customer"
        ],
        "operationId": "Getcustomerbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update customer",
        "tags": [
          "Customer"
        ],
        "operationId": "Updatecustomer",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdatecustomerRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "tags": [
          "Customer"
        ],
        "operationId": "Deletecustomer",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AddadminRequest": {
      "title": "AddadminRequest",
      "example": {
        "username": "admin 8",
        "password": "123456",
        "employeeId": "9e776146-5d61-40bb-9b67-2678446cc26b"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "employeeId": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "password",
        "employeeId"
      ]
    },
    "UpdateadminRequest": {
      "title": "UpdateadminRequest",
      "example": {
        "username": "empladmin",
        "password": "12345678"
      },
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "username",
        "password"
      ]
    },
    "DeleteadminRequest": {
      "title": "DeleteadminRequest",
      "example": {
        "status": 1
      },
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "status"
      ]
    },
...
    "UpdatecustomerRequest": {
      "title": "UpdatecustomerRequest",
      "example": {
        "name": "Emma Stoner",
        "description": "Extra Lorem ipsum dolor sit amet, consectetur adipiscing elit"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "description"
      ]
    },
...
  },
  "security": [
    {
      "token": []
    }
  ],
...
}